fix error & refactor
